yeaaaahhh

+++ db/migrate/20171119192713_create_habbits.rb
@@ -3,7 +3,7 @@ class CreateHabbits < ActiveRecord::Migration[5.1]
     create_table :habbits do |t|
       t.references :country, foreign_key: true
       t.references :system, foreign_key: true
-      t.integer :share, foreign_key: true
+      t.decimal :share, foreign_key: true
       t.text :description
 
       t.timestamps
