+++ db/migrate/20171119193242_create_settings.rb
@@ -2,8 +2,8 @@ class CreateSettings < ActiveRecord::Migration[5.1]
   def change
     create_table :settings do |t|
       t.integer :average_bill
-      t.integer :yandex_net_fee
-      t.integer :yandex_share
+      t.decimal :yandex_net_fee
+      t.decimal :yandex_share
 
       t.timestamps
     end
