+++ db/migrate/20171119193420_create_setups.rb
@@ -3,8 +3,8 @@ class CreateSetups < ActiveRecord::Migration[5.1]
     create_table :setups do |t|
       t.references :website, foreign_key: true
       t.integer :average_bill
-      t.integer :yandex_net_fee
-      t.integer :yandex_share
+      t.decimal :yandex_net_fee
+      t.decimal :yandex_share
 
       t.timestamps
     end
